Add Census I link to Linux Foundation site
This adds a new link to the old "Census I" paper, specifically
to a GitHub repo controlled by the Linux Foundation.
We don't know of a reason the old one would be removed, but it
seems best to maximize the chance of people being able to get
this paper later if they'd like to.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>